// src/store/faqs/FAQs.enum.ts
enum FAQsActionTypeEnum {
    FAQ_LOADING = "FAQ_LOADING",
    FAQ_LOADED = "FAQ_LOADED",
    GET_ALL_FAQ = "GET_ALL_FAQ",
    UPDATE_FAQ = "UPDATE_FAQ",
    GET_FAQ_BY_ID = "GET_FAQ_BY_ID",
    DELETE_FAQ = "DELETE_FAQ",
    ADD_NEW_FAQ = "ADD_NEW_FAQ",
    DO_EMPTY_SINGLE_FAQ = "DO_EMPTY_SINGLE_FAQ"
}

export default FAQsActionTypeEnum;